\docType{package}
\name{bootcorrelations-package}
\alias{bootcorrelations}
\alias{bootcorrelations-package}
\title{A package for bootstrapping differences between mean correlations}
\description{
  This is currently just a test package that I'm using to
  learn about package development
}
\author{
  Jeromy Anglim
}
\seealso{
  The main function in this package is
  \code{\link{boot_mean_cor_diff}}. The summary method can
  then be applied to the returned object.
}

